fix!: correctly set model alias fields
Previously, if a field had an alias, we would set *both* the defined
field name, and the original name from the API.

This was not intentional, only the field defined on the model itself
should be set.

For example, in `ConversationFlowResponse` we defined `api_model_choice
= Field(alias="model_choice")` which resulted in `flow.model_choice` and
`flow.api_model_choice` being accessible.

The latter was not intentional, and `flow.model_choice` will no longer
work after this commit.